ANET-BE-2: Added authentication

+++ API/CityApp-ANET/DTOs/CityDTO.cs
@@ -3,11 +3,10 @@ namespace CityApp_ANET.DTOs
 {
 	public class CityDTO
 	{
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; } = default!;
 
         public string Photo { get; set; } = default!;
     }
 }
-
